pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "space-python-client"
version = "0.1.0"
description = "Python client library for Space - pricing-driven self-adaptation platform"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "ISA-Group" }
]
keywords = ["space", "pricing", "feature-flags", "saas", "websocket"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries"
]
dependencies = [
  "httpx>=0.27.0,<1.0.0",
  "python-socketio>=5.11.0,<6.0.0"
]

[project.optional-dependencies]
redis = [
  "redis>=5.0.0,<6.0.0"
]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0"
]

[project.urls]
Homepage = "https://github.com/Alex-GF/space-python-client"
Documentation = "https://sphere-docs.vercel.app/docs/2.0.1/api/space/SDKs/Backend%20SDKs/space-java-client"
Repository = "https://github.com/Alex-GF/space-python-client"
Issues = "https://github.com/Alex-GF/space-python-client/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/space_client"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra -q"
